// types/Movie.ts
export default interface Movie {
  imdbID: string;
  imdbRating: string;
  imdbVotes: string;
  Title: string;
  Rated: string;
  Released: string;
  Runtime: string;
  Genre: string;
  Director: string;
  Writer: string;
  Actors: string;
  Plot: string;
  Awards: string;
  Poster: string;
  BoxOffice: string;
}